pyproject.toml
[tool.poetry]
name = "scam-sense"
version = "0.1.0"
description = ""
authors = ["akhatre <akhatre@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.2"
pandas = "^2.2.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
tqdm = "^4.66.5"
psycopg2-binary = "^2.9.10"
google-cloud-secret-manager = "^2.21.0"
google-cloud-storage = "^2.18.2"
pandera = "^0.20.4"
xmltodict = "^0.14.2"
xlrd = "^2.0.1"
openpyxl = "^3.1.5"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
django-cors-headers = "^4.6.0"
httpx = "^0.27.2"
sendgrid = "^6.11.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
